Bring in Terrier CI checks (#5)
* Bring in: check-censored, check-clang-tidy, check-lint, check-format, format.

* Modernize FindClangTools from 2013 to 2019 by looking for newer binaries.

* Format fixes.

* Bring in .clang-tidy. Fix lint for everything except [runtime/references] and [whitespace/braces]. The latter is a bug, the former needs discussion.

* Explicitly initialize simd::Vec8's in tests.

* Fix clang-tidy's HeaderFilterRegex whitelist.

* Initial yak-shave at clang-tidy. Need to reason about noexcepts and a malloc.

* Changed to const where possible for lint check

* Extend CPPlint to handle structured binding declarations and various modern checks.

* Desecrate Prashanth's code.

* Update deps.sh.

* Fix last bit of lint.

* Changed const ref in type.h to nonconst ptrs + modified callers of changed functions

* Explicit constructor.

* Different way of ignoring unused member.

* Stupid yaks.

* dev5 g++8 is stupid

* C++ was a mistake

* Yak yak yak.

* Yaaaaak.

* DeadStores is a buggy POS check

* gcc8....

* Another attempt at appeasing g++8

* Appeased g++8

* leave me alone

* First chop

* Missed a few noexcept.

* Removed the wrong noexcepts.

* Replace rand() with C++ rand.

* Disable the test again

* If you assume modern compilers / dev5, we can avoid NOLINTing if constexpr

* Formatter, also you still need to NOLINT else after if constexpr.

* Add tidy-7

* More sacrifices to the lint gods

* only god knows why unique pointers are accused of memory leaks

* bye nolint

* Remove runtime-references check from CPPLINT, clang-tidy.

* Format

* Format and lint.

+++ src/sql/aggregation_hash_table.cpp
@@ -4,14 +4,15 @@ namespace tpl::sql {
 
 AggregationHashTable::AggregationHashTable(util::Region *region,
                                            u32 tuple_size) noexcept
-    : entries_(region, tuple_size), max_fill_(kDefaultInitialTableSize * 0.7) {
+    : entries_(region, tuple_size),
+      max_fill_(static_cast<u64>(kDefaultInitialTableSize * 0.7)) {
   hash_table_.SetSize(kDefaultInitialTableSize);
 }
 
 void AggregationHashTable::Grow() {
   // Resize table
   auto new_size = hash_table_.capacity() * 2;
-  max_fill_ = new_size * 0.7;
+  max_fill_ = static_cast<u64>(new_size * 0.7);
   hash_table_.SetSize(new_size);
 
   // Insert elements again
